Move HTTP post handling tests to requests.test.js
